Add files via upload
The system operates through a Flask-based desktop interface that communicates with the Raspberry Pi via TCP. From the interface, the user selects the operating parameters, the start command, and the terminal. Upon receiving the start command, the Raspberry Pi uses two servo motors and one calibration sensor to take photos around the object according to the selected parameters.

After data acquisition is complete, the images are sent over another TCP communication channel to the desktop, where the Raspberry Pi acts as the master. The images are then reconstructed using the COLMAP library with appropriately tuned parameters. Finally, the result is a PLY file, which can be exported to STL format using MeshLab.